DTFD-MIL/IPD_Brain-main/IPD-Brain-main/requirements.txt
absl-py==1.4.0
addict==2.4.0
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
alabaster==0.7.13
albumentations==1.3.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asciitree==0.3.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
blinker==1.6.2
braceexpand==0.1.7
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.1
comm==0.2.2
contourpy==1.0.7
crc32c==2.3.post0
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dicomweb-client==0.59.1
docutils==0.19
einops==0.6.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
fasteners==0.18
fastjsonschema==2.17.1
fastrlock==0.8.1
filelock==3.10.7
Flask==2.3.2
fonttools==4.39.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2024.3.1
future==0.18.3
geojson==3.1.0
glfw==2.6.2
Glymur==0.12.5
google-auth==2.17.2
google-auth-oauthlib==1.0.0
gpustat==1.1
grpcio==1.53.0
h5py==3.8.0
huggingface-hub==0.22.2
idna==3.4
imagecodecs==2023.3.16
imageio==2.27.0
imagesize==1.4.1
imgaug==0.4.0
imgui==2.0.0
ipykernel==6.27.0
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter-events==0.6.3
jupyter-ydoc==0.2.4
jupyter_client==8.3.0
jupyter_core==5.3.0
jupyter_server==2.7.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.23.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.4
lazy_loader==0.2
lightning-utilities==0.8.0
lit==16.0.0
llvmlite==0.40.1
lmdb==1.4.1
lxml==4.9.3
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.1
monai==1.3.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.0
ninja==1.11.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.57.0
numcodecs==0.11.0
numpy==1.24.2
nvidia-ml-py==11.525.112
nystrom-attention==0.0.11
oauthlib==3.2.2
opencv-python==4.6.0.66
opencv-python-headless==4.8.0.74
openpyxl==3.1.2
openslide-python==1.2.0
overrides==7.3.1
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.10.0
plotly==5.18.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.20.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.11
pydicom==2.3.1
Pygments==2.15.1
pynisher==0.6.4
pynndescent==0.5.10
PyOpenGL==3.1.7
pyparsing==3.0.9
pyperclip==1.8.2
pyrfr==0.9.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==2.0.0
pytorch-toolbelt==0.6.3
pytz==2023.2
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
qudida==0.0.4
referencing==0.29.1
regex==2023.6.3
requests==2.28.2
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.8.10
rsa==4.9
safetensors==0.4.2
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
shapely==2.0.1
SimpleITK==2.2.1
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
sniffio==1.3.0
snowballstemmer==2.2.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
Sphinx==7.0.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.3
sympy==1.11.1
tabulate==0.9.0
tblib==2.0.0
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboard-reducer==0.3.0
tensorboardX==2.6
terminado==0.17.1
threadpoolctl==3.1.0
tiatoolbox==1.4.0
tifffile==2023.3.21
timm==0.9.16
tinycss2==1.2.1
tokenizers==0.15.2
tomli==2.0.1
toolz==0.12.0
topk==1.0
torch==2.0.0
torch-geometric==2.3.1
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.39.3
triton==2.0.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
umap-learn==0.5.3
uri-template==1.3.0
urllib3==1.26.15
visdom==0.2.4
wasabi==1.1.2
wcwidth==0.2.6
webcolors==1.13
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
widgetsnbextension==4.0.10
wrapt==1.15.0
wsidicom==0.9.0
xlrd==2.0.1
y-py==0.5.9
yarl==1.8.2
ypy-websocket==0.8.2
zarr==2.14.2
zict==3.0.0
zipp==3.15.0
